feat: Phase 3 - コンポーネント再編成と統合BaseCard実装
✨ 主な変更内容:
- 統合BaseCardコンポーネントの実装と適用
- domain構造への再編成 (components/domain/{book,bookshelf,memo})
- 重複BaseCardコンポーネントの削除
- 全インポートパスの更新
- 名前付きエクスポート形式への統一

🎯 達成効果:
- コードの重複削除による保守性向上
- 統一されたコンポーネント設計パターン
- ドメイン駆動設計に基づく構造
- TypeScript型安全性の向上

🧪 品質確保:
- 全テスト通過 (165テスト、515アサーション)
- ESLint検証クリア
- ビルド成功確認
