+++ resources/ts/components/domain/book/card/elements/Header.tsx
