+++ resources/ts/components/domain/book/card/elements/Image.tsx
