+++ resources/ts/components/domain/bookshelf/card/DefaultCard.tsx
@@ -1,3 +1,4 @@
+import { BaseCard } from '@/components/common/BaseCard';
 import { Button } from '@/components/common/ui/button';
 import { BookShelfDialogs } from '@/features/bookshelf/components/dialogs/BookShelfDialogs';
 import { useBookShelfState } from '@/features/bookshelf/hooks/useBookShelfState';
@@ -5,7 +6,6 @@ import { BookShelfBase } from '@/types/bookShelf';
 import { Link } from '@inertiajs/react';
 import { Header } from '../elements/Header';
 import { Image } from '../elements/Image';
-import { BaseCard } from './BaseCard';
 
 interface DefaultCardProps extends BookShelfBase {
   image?: string;
@@ -41,13 +41,7 @@ export function DefaultCard({
   const isOwner = type === 'my';
 
   return (
-    <BaseCard
-      variant="card"
-      bookShelfId={bookShelfId}
-      name={name}
-      description={description}
-      isPublic={isPublic}
-    >
+    <BaseCard variant="bookshelf-card">
       <div className="flex flex-row gap-4">
         <Image variant="card" />
