+++ resources/ts/components/domain/bookshelf/elements/Image.tsx
