+++ resources/ts/components/domain/bookshelf/index.tsx
@@ -1,0 +1,10 @@
+// BookShelf domain components exports
+
+// BookShelf card components
+export { DefaultCard as BookShelfDefaultCard } from './card/DefaultCard';
+export { DetailCard as BookShelfDetailCard } from './card/DetailCard';
+
+// BookShelf elements
+export { Header as BookShelfHeader } from './elements/Header';
+export { Image as BookShelfImage } from './elements/Image';
+export { UserInfo as BookShelfUserInfo } from './elements/UserInfo';
