+++ resources/ts/components/domain/index.tsx
@@ -1,0 +1,4 @@
+// Domain components barrel export
+export * from './book';
+export * from './bookshelf';
+export * from './memo';
