+++ resources/ts/components/domain/memo/card/index.tsx
