+++ resources/ts/components/domain/memo/index.tsx
@@ -1,0 +1,2 @@
+// Memo domain components exports
+export { default as MemoCard } from './card/MemoCard';
